Move Ember.TextSupport to ember-views package.
