+++ packages/ember-handlebars/lib/controls/text_area.js
@@ -5,7 +5,7 @@
 */
 import { get } from "ember-metal/property_get";
 import Component from "ember-views/views/component";
-import TextSupport from "ember-handlebars/controls/text_support";
+import TextSupport from "ember-views/mixins/text_support";
 import { observer } from "ember-metal/mixin";
 
 /**
