+++ packages/ember-handlebars/lib/controls/text_field.js
@@ -3,7 +3,7 @@
 @submodule ember-handlebars
 */
 import Component from "ember-views/views/component";
-import TextSupport from "ember-handlebars/controls/text_support";
+import TextSupport from "ember-views/mixins/text_support";
 
 /**
