+++ packages/ember-views/lib/mixins/text_support.js
